backend/requirements.txt
setuptools
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=3.14.0
python-decouple>=3.8
razorpay>=1.5.0
google-generativeai>=0.3.0
Pillow>=10.0.0